move new-message component to forms directory
