+++ resources/js/components/Messages/Message.vue
@@ -119,7 +119,7 @@
 </template>
 
 <script>
-    import NewMessage from './NewMessage'
+    import NewMessage from './../Forms/NewMessage'
     export default {
 
         /**
